改类型

+++ YWFMDB/YWFMDB.h
@@ -21,13 +21,13 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
 @interface YWFMDB : NSObject
 /**
  获取全局的唯一对象
-
+ 
  @return 对象
  */
 + (instancetype)standardYWFMDBDefaults;
 /**
  连接（创建、连接、重连DB）
-
+ 
  @param dbPath 数据库的路径(如:xxx/xxx/yw.db)
  */
 + (void)connectionDB:(NSString *)dbPath;
@@ -37,7 +37,8 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param dbPath 数据库的路径(如:xxx/xxx/yw.db)
  @param key 秘钥
  */
-+ (void)connectionEncryptionDB:(NSString *)dbPath enKey:(NSString *)key;
++ (void)connectionEncryptionDB:(NSString *)dbPath
+                         enKey:(NSString *)key;
 /**
  关闭数据库
  */
@@ -51,7 +52,7 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
 + (BOOL)tableExists:(NSString *)tableName;
 /**
  获取版本号
-
+ 
  @return 版本号
  */
 + (NSString *)version;
@@ -70,7 +71,8 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param isAuto 是否自动根据model的版本号检测表结构是否发生变化(新增字段或者删除字段)(根据需要选择)
  @return 存储成功与否
  */
-+ (BOOL)storageModels:(NSArray<NSObject*>*)model checkTableStructure:(BOOL)isAuto;
++ (BOOL)storageModels:(NSArray<NSObject*>*)model
+  checkTableStructure:(BOOL)isAuto;
 /**
  存储单个model（1、判断表是否存在；2、开始存入数据）(默认检测model的版本号)
  
@@ -85,7 +87,8 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param isAuto 是否自动根据model的版本号检测表结构是否发生变化(新增字段或者删除字段)(根据需要选择)
  @return 存储成功与否
  */
-+ (BOOL)storageModel:(NSObject*)model checkTableStructure:(BOOL)isAuto;
++ (BOOL)storageModel:(NSObject*)model
+ checkTableStructure:(BOOL)isAuto;
 
 /**
  see also：当超多50条记录时，优先删除count最小的数据，在存储当前的数据，参数值：ifExceed=>50，orderBy=>count, desc=>YES
@@ -109,13 +112,16 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param tableName 表名
  @return 存储成功与否
  */
-+ (BOOL)storageModels:(NSArray<NSDictionary*>*)list table:(NSString *)tableName;
++ (BOOL)storageModels:(NSArray<NSDictionary*>*)list
+                table:(NSString *)tableName;
 
 /// 批量存储dict
 /// @param list list的数组
 /// @param tableName 表名
 /// @param version 该表的版本【当该表内部数据结构发现变化，可以增值传入，内部处理】
-+ (BOOL)storageModels:(NSArray<NSDictionary*>*)list table:(NSString *)tableName version:(NSInteger)version;
++ (BOOL)storageModels:(NSArray<NSDictionary*>*)list
+                table:(NSString *)tableName
+              version:(double)version;
 /// 批量存储dict[如果数据有相同，则替换]
 /// @param list list的数组
 /// @param tableName 表名
@@ -124,7 +130,7 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
 + (BOOL)storageModels:(NSArray<NSDictionary*>*)list
                 table:(NSString *)tableName
               mainKey:(NSString *)mainKey
-              version:(NSInteger)version;
+              version:(double)version;
 //MARK: ------------------------------------- 更新 ------------------------------------
 /**
  更新本地存储的model数据
@@ -134,17 +140,22 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param wheres 筛选条件
  @return 更新成功与否
  */
-+ (BOOL)updateWithModel:(NSObject *)model checkTableStructure:(BOOL)automatic where:(NSArray<YWFieldFilter *> *)wheres;
++ (BOOL)updateWithModel:(NSObject *)model
+    checkTableStructure:(BOOL)automatic
+                  where:(NSArray<YWFieldFilter *> *)wheres;
 /**
  指定更新本地存储的model的特定数据
-
+ 
  @param modelClass model的类名
  @param specifiedValue 指定的值
  @param automatic 是否需要自动检测表是否有新增的字段（如果确定没有新增的字典，请传入NO,可以节省一定的开支）(根据需要选择)
  @param wheres 筛选条件
  @return 更新成功与否
  */
-+ (BOOL)updateWithModel:(Class)modelClass specifiedValue:(NSDictionary *)specifiedValue checkTableStructure:(BOOL)automatic where:(NSArray<YWFieldFilter *> *)wheres;
++ (BOOL)updateWithModel:(Class)modelClass
+         specifiedValue:(NSDictionary *)specifiedValue
+    checkTableStructure:(BOOL)automatic
+                  where:(NSArray<YWFieldFilter *> *)wheres;
 //MARK: ------------------------------ 查询操作 ----------------------------------------------
 /**
  查询本地存储的模型对象
@@ -155,12 +166,13 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
 + (NSArray *)queryWithModel:(Class)cls;
 /**
  查询指定条件的本地存储的模型对象
-
+ 
  @param cls model的类型
  @param wheres 筛选条件
  @return 模型对象集合
  */
-+ (NSArray *)queryWithModel:(Class)cls where:(NSArray<YWFieldFilter *> *)wheres;
++ (NSArray *)queryWithModel:(Class)cls
+                      where:(NSArray<YWFieldFilter *> *)wheres;
 /**
  查询指定条件的本地存储的模型对象(结果排序)
  
@@ -168,7 +180,8 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param orders 排序条件
  @return 模型对象集合
  */
-+ (NSArray *)queryWithModel:(Class)cls order:(NSArray<YWFieldOrder *> *)orders;
++ (NSArray *)queryWithModel:(Class)cls
+                      order:(NSArray<YWFieldOrder *> *)orders;
 /**
  分页查询指定条件的本地存储的模型对象
  
@@ -176,7 +189,8 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param page 分页
  @return 模型对象集合
  */
-+ (NSArray *)queryWithModel:(Class)cls limit:(YWPageable *)page;
++ (NSArray *)queryWithModel:(Class)cls
+                      limit:(YWPageable *)page;
 /**
  查询指定条件的本地存储的模型对象,结果并按指定的条件进行排序
  
@@ -185,7 +199,9 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param orders 排序条件
  @return 模型对象集合
  */
-+ (NSArray *)queryWithModel:(Class)cls where:(NSArray<YWFieldFilter *> *)wheres order:(NSArray<YWFieldOrder *> *)orders;
++ (NSArray *)queryWithModel:(Class)cls
+                      where:(NSArray<YWFieldFilter *> *)wheres
+                      order:(NSArray<YWFieldOrder *> *)orders;
 /**
  分页查询指定条件的本地存储的模型对象,结果并按指定的条件进行排序
  
@@ -194,7 +210,9 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param page 分页
  @return 模型对象集合
  */
-+ (NSArray *)queryWithModel:(Class)cls where:(NSArray<YWFieldFilter *> *)wheres limit:(YWPageable *)page;
++ (NSArray *)queryWithModel:(Class)cls
+                      where:(NSArray<YWFieldFilter *> *)wheres
+                      limit:(YWPageable *)page;
 /**
  分页查询本地存储的模型对象,结果并按指定的条件进行排序
  
@@ -203,7 +221,9 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param page 分页
  @return 模型对象集合
  */
-+ (NSArray *)queryWithModel:(Class)cls order:(NSArray<YWFieldOrder *> *)orders limit:(YWPageable *)page;
++ (NSArray *)queryWithModel:(Class)cls
+                      order:(NSArray<YWFieldOrder *> *)orders
+                      limit:(YWPageable *)page;
 /**
  分页查询指定条件的本地存储的模型对象,结果并按指定的条件进行排序
  
@@ -213,7 +233,10 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param page 分页
  @return 模型对象集合
  */
-+ (NSArray *)queryWithModel:(Class)cls where:(NSArray<YWFieldFilter *> *)wheres order:(NSArray<YWFieldOrder *> *)orders limit:(YWPageable *)page;
++ (NSArray *)queryWithModel:(Class)cls
+                      where:(NSArray<YWFieldFilter *> *)wheres
+                      order:(NSArray<YWFieldOrder *> *)orders
+                      limit:(YWPageable *)page;
 /**
  分页查询指定条件的本地存储的数据,结果并按指定的条件进行排序【返回结果无需转成对象】
  
@@ -224,10 +247,14 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param page 分页
  @return 模型对象集合
  */
-+ (NSArray *)queryWithTableName:(NSString *)tableName fields:(NSDictionary *)fields where:(NSArray<YWFieldFilter *> *)wheres order:(NSArray<YWFieldOrder *> *)orders limit:(YWPageable *)page;
++ (NSArray *)queryWithTableName:(NSString *)tableName
+                         fields:(NSDictionary *)fields
+                          where:(NSArray<YWFieldFilter *> *)wheres
+                          order:(NSArray<YWFieldOrder *> *)orders
+                          limit:(YWPageable *)page;
 /**
  函数查询
-
+ 
  @param cls model的类型
  @param function 函数条件
  @return 具体结果
@@ -251,7 +278,7 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
 + (BOOL)deleteTableWithModel:(Class)cls;
 /**
  删除该表的所有数据
-
+ 
  @param table 表名
  @return 删除成功与否
  */
@@ -263,7 +290,8 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param wheres 筛选条件
  @return 删除成功与否
  */
-+ (BOOL)deleteTableWithModel:(Class)cls where:(NSArray<YWFieldFilter *> *)wheres;
++ (BOOL)deleteTableWithModel:(Class)cls
+                       where:(NSArray<YWFieldFilter *> *)wheres;
 /**
  删除该表指定条件的数据
  
@@ -271,7 +299,8 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param wheres 筛选条件
  @return 删除成功与否
  */
-+ (BOOL)deleteTable:(NSString *)table where:(NSArray<YWFieldFilter *> *)wheres;
++ (BOOL)deleteTable:(NSString *)table
+              where:(NSArray<YWFieldFilter *> *)wheres;
 
 //MARK: ------------------------------------- 导出（CSV/TXT） ------------------------------------
 /**
@@ -281,7 +310,8 @@ typedef NS_ENUM(NSInteger, YWFMDBExportType) {
  @param type csv/txt
  @return 导出的路径
  */
-+ (NSString *)exportTable:(NSString *)table type:(YWFMDBExportType)type;
++ (NSString *)exportTable:(NSString *)table
+                     type:(YWFMDBExportType)type;
 
 @end
