+++ YWFMDB/YWFMDB.m
@@ -241,7 +241,7 @@ static YWFMDB *singletonInstance = nil;
 /// @param version 该表的版本【当该表内部数据结构发现变化，可以增值传入，内部处理】
 + (BOOL)storageModels:(NSArray<NSDictionary*>*)list
                 table:(NSString *)tableName
-              version:(NSInteger)version{
+              version:(double)version{
     return [self storageModels:list table:tableName mainKey:nil version:version];
 }
 /// 批量存储dict[如果数据有相同，则替换]
@@ -252,7 +252,7 @@ static YWFMDB *singletonInstance = nil;
 + (BOOL)storageModels:(NSArray<NSDictionary*>*)list
                 table:(NSString *)tableName
               mainKey:(NSString *)mainKey
-              version:(NSInteger)version{
+              version:(double)version{
     if (!singletonInstance.queue) {
         [self log:@"请先连接数据库"];
         return NO;
